Add some component test for Analyze controller

+++ src/main/java/pl/com/sremski/domain/Analysis.java
@@ -6,6 +6,6 @@ import lombok.Setter;
 public class Analysis {
 
     @Setter @Getter private String analyseDate;
-    @Setter @Getter private long analyseTime;
+    @Setter @Getter private Double analyseTime;
     @Setter @Getter private Details details;
 }
